fix league naming
