--- store/readonly.js
@@ -1,11 +1,0 @@
-var knex = require('knex')(
-{
-    client: 'pg',
-    connection: "readonly:readonly@localhost/yasp",
-    pool:
-    {
-        min: 1,
-        max: 5,
-    },
-});
-module.exports = knex;
